Remove exceptionsmanager from android, Use only module
